Allow coercion of a dict into a Peer
When passing the JSON exported data back into Peer or Server, the provided list of dict
would raise a ValueError. This allows passing a list of dict as the value for the `peers`
keyword argument.

Additionally, this fixes a bug with pre/post up/down rules which were being incorrectly
appended to the internal list, instead of extending the list when the value was iterable.

ref: #10, #12